remove use of \u2000 as an alternate space character

+++ ytdownloader.py
@@ -13,6 +13,7 @@ from shutil import which
 from pathlib import Path
 import tkinter as tk
 import tkinter.messagebox
+from tkinter import simpledialog
 import tkinter.font as tkFont
 import tkinter.ttk as ttk
 from tkinterdnd2 import *
@@ -23,6 +24,9 @@ YTDL_DOWNLOAD_DIR = expanduser("~") + "/Music/ytdl"
 STAGING_DIR = YTDL_DOWNLOAD_DIR + "/staging"
 YTDL_PATH = None
 
+# naming fixes:
+# Folk Alley Sessions: Anna Egge Girls..
+# Cristina Vane Getting High in Hotel Rooms
 
 def logit(msg):
     timestr = datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S: ")
@@ -91,67 +95,82 @@ def schedule_check(command_thread):
 
 def clean_filepath(filepath):
     FIELD_SEPARATOR='^'
-    new_file = filepath
+    new_name = os.path.basename(filepath)
+
+    if not (filepath.endswith('.wav') or filepath.endswith(".mp3")):
+        return filepath
 
     # remove parenthetical and bracketed text
-    new_file = re.sub(r"[\(\[].*?[\)\]]", "", new_file)
-    new_file = re.sub(r'- \d+ -', FIELD_SEPARATOR, new_file)
+    new_name = re.sub(r"[\(\[\{].*?[\)\]\}]", "", new_name)
+    new_name = re.sub(r'- \d+ -', FIELD_SEPARATOR, new_name)
 
     # replace quoted song with seperator, e.g. John Craige "Judias"
-#    match = re.search(r"([^'\"]*)['\"]([^'\"]*)", new_file)
+#    match = re.search(r"([^'\"]*)['\"]([^'\"]*)", new_name)
 #    if match and len(match.groups()) == 2:
-#        new_file = f"{match.group(1)} {FIELD_SEPARATOR} {match.group(2)}"
+#        new_name = f"{match.group(1)} {FIELD_SEPARATOR} {match.group(2)}"
 
     # replace quoted song with seperator, e.g. John Craige "Judias"
     WIERD_QUOTE = '＂'
-    if new_file.find(WIERD_QUOTE) > 0:
-        new_file = new_file.replace(WIERD_QUOTE, FIELD_SEPARATOR, 1)
-        new_file = new_file.replace(WIERD_QUOTE, '', 1)
+    if new_name.find(WIERD_QUOTE) > 0:
+        new_name = new_name.replace(WIERD_QUOTE, FIELD_SEPARATOR, 1)
+        new_name = new_name.replace(WIERD_QUOTE, '', 1)
 
-    if new_file.find('Official Track') >= 0:
-        new_file = new_file.replace('Official Track', '')
+    if new_name.find('Official Track') >= 0:
+        new_name = new_name.replace('Official Track', '')
 
-    if new_file.find('OFFICIAL MUSIC VIDEO') >= 0:
-        new_file = new_file.replace('OFFICIAL MUSIC VIDEO', '')
+    if new_name.find('Official Lyric Video') >= 0:
+        new_name = new_name.replace('Official Lyric Video', '')
 
-    if new_file.find('NA_') >= 0:
-        new_file = new_file.replace('NA_', '')
+    if new_name.find('Lyric Video') >= 0:
+        new_name = new_name.replace('Lyric Video', '')
 
-    if new_file.find('｜') >= 0:
-        new_file = new_file.replace('｜', FIELD_SEPARATOR)
+    if new_name.find('OFFICIAL MUSIC VIDEO') >= 0:
+        new_name = new_name.replace('OFFICIAL MUSIC VIDEO', '')
 
-    if new_file.find(' : ') >= 0:
-        new_file = new_file.replace(' : ', FIELD_SEPARATOR)
+    if new_name.find('NA_') >= 0:
+        new_name = new_name.replace('NA_', '')
 
-    if new_file.find('＂') >= 0:  # special fat double quote from &quot; in html
-        new_file = new_file.replace('＂', '')
+    if new_name.find('｜') >= 0:
+        new_name = new_name.replace('｜', FIELD_SEPARATOR)
 
-    if new_file.find('"') >= 0:  # regular double quote
-        new_file = new_file.replace('"', '')
+    if new_name.find(' : ') >= 0:
+        new_name = new_name.replace(' : ', FIELD_SEPARATOR)
 
-    if new_file.find('-') >= 0:
-        new_file = new_file.replace('-', FIELD_SEPARATOR)
+    if new_name.find('＂') >= 0:  # special fat double quote from &quot; in html
+        new_name = new_name.replace('＂', '')
 
-    if new_file.find('_') >= 0:
-        new_file = new_file.replace('_', ' ' + FIELD_SEPARATOR + ' ')
+    if new_name.find('"') >= 0:  # regular double quote
+        new_name = new_name.replace('"', '')
 
-    if new_file.find('–') >= 0:
-        new_file = new_file.replace('–', FIELD_SEPARATOR)
+    if new_name.find('-') >= 0:
+        new_name = new_name.replace('-', FIELD_SEPARATOR)
 
-    if new_file.find('Official HD Audio') >= 0:  # regular double quote
-        new_file = new_file.replace(' Official HD Audio', '')
+    if new_name.find('_') >= 0:
+        new_name = new_name.replace('_', ' ' + FIELD_SEPARATOR + ' ')
 
-    if new_file.find('Official Music Video') >= 0:  # regular double quote
-        new_file = new_file.replace(' Official Music Video', '')
+    if new_name.find('–') >= 0:
+        new_name = new_name.replace('–', FIELD_SEPARATOR)
 
-    # just in case we lost the suffix in the transform.
-#    if not new_file.endswith('.wav'):
-#        new_file += '.wav'
+    if new_name.find('Official HD Audio') >= 0:  # regular double quote
+        new_name = new_name.replace(' Official HD Audio', '')
+
+    if new_name.find('Official Music Video') >= 0:  # regular double quote
+        new_name = new_name.replace(' Official Music Video', '')
+
+    if new_name.find(f"{FIELD_SEPARATOR} {FIELD_SEPARATOR}") >= 0:
+        new_name = new_name.replace(f"{FIELD_SEPARATOR} {FIELD_SEPARATOR}", FIELD_SEPARATOR)
+
+    if new_name.find(f"{FIELD_SEPARATOR} .") >= 0:
+        new_name = new_name.replace(f"{FIELD_SEPARATOR} .", ".")
+
+    splitAr = new_name.split(FIELD_SEPARATOR)
+    if len(splitAr) != 2:
+        new_name  = simpledialog.askstring("File Name", "Track Name\t\t\t\t\t\t", initialvalue=f"{new_name}")
+    new_file = f"{os.path.dirname(filepath)}/{new_name}"
 
     # trim secondary artist names, e.g. anything after a comma
     nameAr = new_file.split(FIELD_SEPARATOR)
     commaIdx = nameAr[0].find(',')
-    print(f"tp: {nameAr[0]}, {commaIdx}")
     if commaIdx > 0 and len(nameAr) > 1:
         new_file = f"{nameAr[0][0:commaIdx]} {FIELD_SEPARATOR} {nameAr[1]}"
 
@@ -169,28 +188,20 @@ def check_if_done(command_thread):
     else:
         root.bell()
         errmsg = str(command_thread.stderr) # returns '\\\' even when there is no error.
-        length = len(errmsg)
-
-        fatalMsg = errmsg.find("HTTP Error 403 Forbidden") > 0
-        ignoreMsg = ((errmsg.find('Skipping player responses from android') > 0) | \
-                     (errmsg.find('You may experience throttling for some formats') > 0) | \
-                     (errmsg.find('Signature extraction failed:') > 0))
-        print("msg: {}, {}".format(ignoreMsg, fatalMsg))
-        ignoreMsg = (ignoreMsg == True) and (fatalMsg == False)
-
         if errmsg.find('File name too long') > 0:
             tk.messagebox.showwarning(title='Error', message='Artist name too long. Click Okay to download using UNKNOWN for the artist name')
             control_panel._fetch_url(False)
             return
 
-        if (command_thread.process.returncode == 0 and (fatalMsg == False) and (ignoreMsg or len(errmsg) < 4)):
+        if command_thread.process.returncode == 0:
             res = str(command_thread.stdout, 'utf-8')
             idx1 = res.rfind("Destination: ") + 13
             idx2 = res.find(".wav", idx1)
             if idx1 > 13 and idx2 > idx1:
+                errmsg = ''
                 filepath =  res[idx1:idx2+4]
+                logit("Downloaded file: " + filepath)
                 filepath = clean_filepath(filepath)
-                logit("Add file: " + filepath)
                 trim_audio(filepath)
 
                 listbox.populate_list()
@@ -198,30 +209,13 @@ def check_if_done(command_thread):
                 control_panel.url.config(cursor="")
                 #control_panel.url.config({"background": "Green"})
                 control_panel.url.update()
-            else:
-                errmsg = res
 
-        if len(errmsg) > 4 and not ignoreMsg:
+        if len(errmsg) > 0:
+            control_panel.url.config(cursor="")
             tk.messagebox.showwarning(title='Error', message=errmsg)
             #control_panel.url.config({"background": "Red"})
             #control_panel.url.update()
 
-
-def execute_command(cmd):
-    try:
-        #logit("Execute: +{}+\n".format(cmd))
-        p = subprocess.Popen(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True)
-        (output, err) = p.communicate()
-        p_status = p.wait()
-        root.bell()
-        if p_status != 0:
-            msg = "Returned +{}+, +{}+".format(output, str(err))
-            tk.messagebox.showwarning(title='Error', message=msg)
-
-        return p_status == 0
-    except Exception as ioe:
-        logit('Exception executing command: {}, {}'.format(cmd, ioe))
-        return False
 
 def delete_file_after_delay(file_name):
     if os.path.exists(file_name):
@@ -229,7 +223,7 @@ def delete_file_after_delay(file_name):
         #os.remove(file_name)
 
 class ControlPanel(object):
-    def __init__(self, frame):
+    def __init__(self, root):
         self.player = None
         self.list_widget = None
 
@@ -237,17 +231,17 @@ class ControlPanel(object):
         lbl = tk.Label(master=top_frame, text='URL:')
         lbl.place(x=0, y=1)
         self.urlEntry = tk.StringVar()
-        self.url = tk.Entry(master=top_frame, textvariable=self.urlEntry, width=40)
+        self.url = tk.Entry(master=top_frame, textvariable=self.urlEntry, width=39)
         self.url.bind('<Return>', self._on_url_enter)
         self.url.place(x=30, y=0)
         button = tk.Button(master=top_frame, command= self._fetch_url, text="Fetch", width=5, fg="black", )
-        button.place(x=370, y=0)
+        button.place(x=398, y=4)
         button = tk.Button(master=top_frame, command= self._reload, text="Reload", width=5, fg="black", )
-        button.place(x=450, y=0)
+        button.place(x=450, y=4)
         button = tk.Button(master=top_frame, command= self._play_file, text="Play", width=5, fg="black", )
-        button.place(x=530, y=0)
+        button.place(x=500, y=4)
         button = tk.Button(master=top_frame, command= self._stop_play, text="Stop", width=5, fg="black", )
-        button.place(x=610, y=0)
+        button.place(x=550, y=4)
         top_frame.pack(fill=tk.X)
 
 
@@ -283,7 +277,8 @@ class ControlPanel(object):
     def delete_click(self, event):
         item = self.list_widget.tree.selection()[0]
         print("delete: {}".format(item))
-        os.remove(item)
+        if os.path.exists(item):
+            os.remove(item)
         self._reload()
 
 
@@ -312,18 +307,74 @@ class ControlPanel(object):
             schedule_check(download_thread)
 
 class FilePickerListbox(object):
-
     def __init__(self, frame):
-        self.table_header = ['File']
+        self.have_shift = False
         self.tree = None
         self.item_name = None
         self._setup_widgets(frame)
-        self._build_tree()
         self._set_file_header()
 
         self.tree.drag_source_register(1, DND_FILES)
         self.tree.dnd_bind('<<DragInitCmd>>', self.drag_init)
         self.tree.dnd_bind('<<DragEndCmd>>', self.drag_end)
+        self.tree.bind("<<TreeviewSelect>>", self.on_item_select)
+        self.tree.bind("<Button-1>", self.on_tree_click)
+
+        self.tree.bind("<Shift-Up>", lambda e: self.on_shift_arrow(e, "up"))
+        self.tree.bind("<Shift-Down>", lambda e: self.on_shift_arrow(e, "down"))
+
+    def on_tree_click(self, event):
+        """Handle clicks for checkbox toggling and row selection."""
+        region = self.tree.identify_region(event.x, event.y)
+        column = self.tree.identify_column(event.x)
+        item = self.tree.identify_row(event.y)
+
+        if region == "cell" and column == "#1":
+            self.move_to_staging_dir(item)
+            self.tree.delete(item)
+            return "break" # prevent propagation
+
+
+
+    def on_shift_arrow(self, event, direction):
+        selection = self.tree.selection()
+        items = self.tree.get_children("")
+        
+        if not selection:
+            # if nothing selected, start at first/last
+            idx = 0 if direction == "down" else len(items) - 1
+        else:   
+            # last focused item index
+            focus = self.tree.focus() or selection[-1]
+            try:
+                idx = items.index(focus)
+            except ValueError: 
+                idx = 0
+    
+            idx = max(0, min(len(items) - 1, idx + (1 if direction == "down" else -1)))
+        
+        new_item = items[idx]
+    
+        # Add new item to selection
+        self.tree.selection_add(new_item)
+        self.tree.focus(new_item) 
+        self.tree.see(new_item)
+        return "break"  # prevent default move
+
+
+    def on_shift_change(self, event):
+        if event.state == 1:
+            self.have_shift = True
+        else:
+            self.have_shift = False
+
+        print(f"Shift change {self.have_shift}")
+
+    def on_item_select(self, event):
+        print(f"Shift {self.have_shift}")
+        selected_items = self.tree.selection() # Get the selected item(s)
+        for item in selected_items:
+            print("Selected item:", self.tree.item(item, "text"))
 
     def _set_file_header(self):
         msg = 'Files: ({})'.format(len(self.tree.get_children()))
@@ -335,41 +386,57 @@ class FilePickerListbox(object):
         data = ()
         self.item_name = self.tree.selection()
         if self.item_name:
-            print("doing drag: {}".format(self.item_name))
-            path, name = os.path.split(self.item_name[0])
-            if name.find("LID_") == 0:
-                stage_file = self.item_name
-            else:
-                # convert spaces to En Quad space because DnD uses space as a file seperator
-                name_safe = name.replace(' ', '\u2000')
-                stage_file = STAGING_DIR + "/" + name_safe
-                print("move {} -> {}".format(self.item_name[0], stage_file))
-                os.rename(self.item_name[0], stage_file)
+            stage_files = []
+            for path in self.item_name:
+                stage_file = self.move_to_staging_dir(path)
+                stage_files.append(stage_file)
 
             self.tree.dragging = True
-            return ((COPY, MOVE), (DND_FILES), stage_file)
+            return ((COPY, MOVE), (DND_FILES), stage_files)
         else:
             return ((), (), ()) # what to return here?
+
+    def move_to_staging_dir(self, file_path):
+        if file_path.find("/LID_") > 0:
+            return file_path
+
+        path, name = os.path.split(file_path)
+        stage_file = STAGING_DIR + "/" + name
+        os.rename(file_path, stage_file)
+        return stage_file
 
     def drag_end(self, event):
         print("drag end")
         action = event.action
         # reset the "dragging" flag to enable drops again
-        file_name = self.item_name[0]
-        print("drag end: {}, {}".format(action, file_name))
-        # Don't remove LID files since they are reused.
-        if file_name.find("/LID_") < 0:
-            self.tree.delete(file_name)
+        for item in self.item_name:
+            file_path = item
+
+            # Don't remove LID files since they are reused.
+            if file_path.find("/LID_") < 0:
+                if action == 'refuse_drop':
+                    path, file_name = os.path.split(item)
+                    #stage_file = STAGING_DIR + "/" + file_name.replace(' ', '\u2000')
+                    stage_file = STAGING_DIR + "/" + file_name
+                    dest_file = path + "/" + file_name
+                    os.rename(stage_file, dest_file)
+                else:
+                    self.tree.delete(file_path)
 
         self.tree.dragging = False
-
 
     def _setup_widgets(self, frame):
         container = ttk.Frame(frame)
         container.pack(fill='both', expand=True)
 
         # create a treeview with dual scrollbars
-        self.tree = ttk.Treeview(columns=self.table_header, height=520, show="headings")
+        self.tree = ttk.Treeview(columns=('Select', 'File'), height=520, show="headings")
+        self.tree.heading('Select', text='#')
+        self.tree.column('Select',  width=25, anchor="center", stretch=False)
+
+        self.tree.heading('File', text='File')
+        self.tree.column("File", anchor="w", stretch=True)
+
         vsb = ttk.Scrollbar(orient="vertical", command=self.tree.yview)
         hsb = ttk.Scrollbar(orient="horizontal", command=self.tree.xview)
         self.tree.configure(yscrollcommand=vsb.set, xscrollcommand=hsb.set)
@@ -378,6 +445,8 @@ class FilePickerListbox(object):
         hsb.grid(column=0, row=1, sticky='ew', in_=container)
         container.grid_columnconfigure(0, weight=1)
         container.grid_rowconfigure(0, weight=1)
+
+        self.populate_list()
 
     def reload_list(self):
         self.tree.delete(*self.tree.get_children())
@@ -392,6 +461,7 @@ class FilePickerListbox(object):
         files = glob.glob(mpe_path + ".mp3") + glob.glob(mpe_path + ".wav")
         for mpefile in files:
             print("move file: " + mpefile)
+            trim_audio(mpefile)
             shutil.move(mpefile, YTDL_DOWNLOAD_DIR)
 
         prefix_len = len(ytdl_path) - 1
@@ -401,7 +471,6 @@ class FilePickerListbox(object):
         mergeFiles = []
         for filepath in files:
             filepath = clean_filepath(filepath)
-            trim_audio(filepath)
             name = filepath[prefix_len: len(filepath)]
             mergeFiles.append([name, filepath])
 
@@ -418,43 +487,41 @@ class FilePickerListbox(object):
             if not name in loadedFiles:
                 loadedFiles.append(name)
                 #print("add: " + name)
-                self.tree.insert('', 'end', iid=fileAr[1], text=name, values=([name]))
+                self.tree.insert('', 'end', iid=fileAr[1], text=name, values=(['x', name]))
             else:
                 pass #logit("File already exists: " + name)
 
         self._set_file_header()
 
 
-    def _build_tree(self):
-        for col in self.table_header:
-            self.tree.heading(col, text=col.title(), command=lambda c=col: sortby(self.tree, c, 0))
-            # adjust the column's width to the header string
-            self.tree.column(col, width=tkFont.Font().measure(col.title()))
-
-        self.populate_list()
 
 def resize(event):
     logit("height: ", event.height, "width: ", event.width)
 
-if __name__ == '__main__':
-    root = TkinterDnD.Tk()
-    #root = tk.Tk()
-    root.title("Youtube Download Tool")
-    root.geometry("560x490")
-    control_panel = ControlPanel(root)
+def create_app(root):
+    global YTDL_PATH, control_panel, listbox
 
     YTDL_PATH = shutil.which('yt-dlp')
-    #YTDL_PATH = "/Users/Barbara/src/youtube-dl/youtube-dl"
-
-    logit("ytdl path: {}".format(YTDL_PATH))
-
     if not os.path.exists(YTDL_DOWNLOAD_DIR):
         os.makedirs(YTDL_DOWNLOAD_DIR)
 
+    logit("ytdl path: {}".format(YTDL_PATH))
+
+    if not root:
+        root = TkinterDnD.Tk()
+        root.title("Youtube Download Tool")
+        root.geometry("560x490")
+
+
+
+    control_panel = ControlPanel(root)
     list_frame = tk.Frame(master=root, height=300)
     list_frame.pack(fill=tk.X)
     listbox = FilePickerListbox(list_frame)
     control_panel.set_list_widget(listbox)
-    #root.bind("<Configure>", resize)
+    return root
+
+if __name__ == '__main__':
+    root = create_app(None)
     root.mainloop()
